requirements.txt
boto3>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
psycopg[binary]>=3.1.12
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
unstructured>=0.10.8
pdf2image>=1.16.3
pdfminer.six>=20221105
pubget
biopython>=1.81
pgvector>=0.2.4